+++ numismatic/exchanges/__init__.py
@@ -1,0 +1,4 @@
+from numismatic.exchanges.base import Exchange
+from numismatic.exchanges.bitfinex import BitfinexExchange
+from numismatic.exchanges.luno import LunoExchange
+__all__ = ["Exchange", "BitfinexExchange", "LunoExchange", "luno"]
