+++ numismatic/feeds/__init__.py
@@ -1,0 +1,2 @@
+from numismatic.feeds.base import Feed
+__all__ = ["Feed", "cryptocompare", "luno"]
